Added gradle build, tests and Travis build.
Test framework uses a bash script to exercise gw.  It assumes that at least "gradle" is
installed and that "gradlew" is not installed in a parent path to the project build.

The tests should be run once before and once after a gradle wrapper has been installed.
Refer to .travis.yml for details.

As of this commit the test will fail.