Don't change-directory to the nearest parent with a build.gradle.  Fixed #22.
gw should match the behaviour of both "gradle" and "gradlew" without surprises.
Prior to this change, if there is a subproject without its own build.gradle, gw
will change-up to some ancestor project which does have a build.gradle. This results
in different behaviour to "gradle" and "gradlew" when executed from the same
location.

This change removes the cd.  A proper gradle multi-project structure will
still allow the resulting gradle runtime to locate the root project and its
build.gradle.